🌱 updated seed module
