+++ libs/api/database/seed/src/lib/seed.module.ts
@@ -6,5 +6,3 @@ import { Module } from '@nestjs/common';
   exports: []
 })
 export class SeedModule {}
-
-// TODO: 🗑️ remove this module
