+++ libs/api/database/seed/src/lib/seeds/index.ts
@@ -1,3 +1,3 @@
 export * from './tenure';
 export * from './location';
-export * from './user';
+export * from './person';
