+++ libs/api/database/seed/src/lib/seeds/person/index.ts
@@ -3,5 +3,5 @@ export * from './demographics';
 export * from './identification';
 export * from './social-connection';
 
-export * from './usama.user';
-export * from './eram.user';
+export * from './usama.person';
+export * from './eram.person';
